backend/code/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
django-cors-headers
djangorestframework_jwt
decorator==4.3.0
Django==2.1.3
djangorestframework==3.9.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
kombu==4.2.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.0
pytest==4.0.1
pytest-django==3.4.4
pytz==2018.7
facebook-sdk
django-bootstrap3
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7